% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD.O.3.histo}
\alias{ICD.O.3.histo}
\title{ICD-O-3 histological codes & grouppings}
\format{A data frame with 4 variables:
\describe{
\item{\code{histology}}{}
\item{\code{histology.description}}{}
\item{\code{histology.behavior}}{}
\item{\code{histology.behavior.description}}{}
}

For further details, see \url{http://seer.cancer.gov/icd-o-3/} and \url{http://codes.iarc.fr/topography}}
\usage{
ICD.O.3.histo
}
\description{
The ICD-O-3 histological codes & grouppings as used by ICD-O-3 SEER Site/Histology Validation List
}
\keyword{datasets}

