% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD.O.3.topo}
\alias{ICD.O.3.topo}
\title{ICD-O-3 topographical codes & grouppings}
\format{A data frame with six variables:
\describe{
\item{\code{ICDO3.code}}{ICD-O-3 topographical code (without the decimal point)}
\item{\code{locality}}{ICD-O-3 locality}
\item{\code{ICDO3.topography.group}}{ICD-O-3 groupping}
\item{\code{SEER.site.recode}}{SEER group}
\item{\code{SEER.site.description}}{SEER group description}
\item{\code{SEER.site.recode.id}}{SEER group id as used by \code{ICD.O.3.hist}}
}

For further details, see \url{http://seer.cancer.gov/icd-o-3/} and \url{http://codes.iarc.fr/topography}}
\usage{
ICD.O.3.topo
}
\description{
The ICD-O-3 locality codes & grouppings as used by ICD-O-3 SEER Site/Histology Validation List
}
\keyword{datasets}

