% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD.O.3.valid.histotopo}
\alias{ICD.O.3.valid.histotopo}
\title{ICD-O-3 valid mappings between histological & topographical codes}
\format{A data frame with 2 variables:
\describe{
\item{\code{histology.behavior}}{maps to \code{\link{ICD.O.3.histo}}}
\item{\code{SEER.site.recode.id}}{maps to \code{\link{ICD.O.3.topo}}}
}

For further details, see \url{http://seer.cancer.gov/icd-o-3/} and \url{http://codes.iarc.fr/topography}}
\usage{
ICD.O.3.valid.histotopo
}
\description{
The ICD-O-3 valid histological & topographical pairs as provided by ICD-O-3 SEER Site/Histology Validation List
}
\keyword{datasets}

