% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{load.SEER.data.raw}
\alias{load.SEER.data.raw}
\title{SEER incidence txt data loader}
\usage{
load.SEER.data.raw(file, n_max = -1)
}
\arguments{
\item{file}{a path to data file, e.g. \code{.../SEER_1973_2013_TEXTDATA/incidence/yr1973_2013.seer9/URINARY.TXT}}

\item{n_max}{maximal number of records to read (\code{-1} - no limit)}
}
\description{
Loads the incidence data from SEER base to raw (textual) format
}

